Moved all the functions to calculate distance into a new service.
